move test to a better place
